Cleaned up some code.
This is a test build that attempts to use the selected opponent's pane to decide who to ask a card from. It is a modification of Chad's original code. It is by no means complete, but it does allow to test in a 3 player game. There is the larger issue of making sure the right opponents (player0,1,2 etc) are in the right windows.